Add files via upload
These are the latest test files. pi_socket_code.py is run on the raspberry pi, and attempts to establish three connections with the ev3s. ev3_socket_code sets up a listening server from the ev3, ready to receive the pi's connections. Once connection is established, an infinite loop of user input for commands begins, sending them off to each of the ev3s and waiting for each to acknowledge the command before continuing.